+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.0.3] - 2020-05-18
+ 
+* Adding Point Distance
+* Adding AutoEditMode option
+* Adding PointDistance option
+
 ## [0.0.2+2] - 2020-05-15
  
 * Fixing Readme Error